Add English translation for values/strings.xml
Given that a significant portion the messages shown in the application
are country-dependent and must set in accordance with the local health
authority policies, this is just intended to be used as a template to
help translators, developers and testers.

It may be convenient to make this the default, and move
values/strings.xml to values-cz/strings.xml, but that's a decision for
the project maintainers.